Relink the OpenBLAS library with R

The link_again function relinks the OpenBLAS library with the R language, which is useful for correcting unlinking problems that are common when the operating system is updated.

link_again(restart_r = TRUE)

Arguments

restart_r

If TRUE (default), a new R session is started after linking the OpenBLAS library.

Details

The link_again function can relink the R language with the OpenBLAS library. Thus, link_again will only perform the relinking when the ropenblas function has been used in a previous R session for the initial binding of the R language with the OpenBLAS library.

Relinking is useful in situations of operating system updates. In some updates, it is possible that the OpenBLAS library compiled in the /opt directory gets unlinked. In this scenario, when the OpenBLAS library has already been compiled using the ropenblas function, the link_again function performs a new link without the need to recompile, thus making the process less time consuming.

Note

In situations where there was a disconnection due to an operating system update, the ropenblas function can be used to re-link the OpenBLAS library with the R language, however, it will be necessary to compile the OpenBLAS library again. If you are interested in recompiling the OpenBLAS library and linking with R, use the ropenblas function. If the interest is to take advantage of a previous compilation of the OpenBLAS library, the link_again function may be useful.

See also

Examples

# link_again()