Download, Compile and Link OpenBLAS Library with R

Links R with an optimized version of the BLAS library (OpenBLAS).

ropenblas(x = NULL, restart_r = TRUE)

Arguments

x

OpenBLAS library version to be considered. By default, x = NULL.

restart_r

If TRUE, a new R session is started after compiling and linking the OpenBLAS library.

Value

Returns a message informing you if the procedure occurred correctly. You will also be able to receive information about missing dependencies.

Details

The ropenblas() function will only work on Linux systems. When calling the ropenblas() function on Windows, no settings will be made. Only a warning message will be issued informing you that the configuration can only be performed on Linux systems.

The function will automatically download the latest version of the OpenBLAS library. However, it is possible to inform old versions to the single argument of ropenblas(). The ropenblas() function downloads, compiles and links R to use the OpenBLAS library. Everything is done very simply, just by loading the library and invoking the function ropenblas().

Considering using the OpenBLAS library rather than BLAS may bring extra optimizations for your code and improved computational performance for your simulations, since OpenBLAS is an optimized implementation of the BLAS library.

You must install the following dependencies on your operating system (Linux):

GNU Make;

GNU GCC Compiler (C and Fortran).

Your Linux operating system may already be configured to use the OpenBLAS library. Therefore, R will most likely already be linked to this library. To find out if R is using the OpenBLAS library, in R, do:

extSoftVersion()["BLAS"]

If R is using the OpenBLAS library, something like /any_directory/libopenblas.so should be returned. Therefore, the name 'openblas' should be in the returned shared object (file with a .so extension).

If the ropenblas() function can identify that R is using the version of OpenBLAS you wish to configure, a warning message will be returned asking if you would really like to proceed with the configuration again.

The ropenblas() function will download the desired version of the OpenBLAS library, compile and install it in the /opt directory of your operating system. If the directory does not exist, it will be created so that the installation can be completed. Subsequently, files from the version of BLAS used in R will be symbolically linked to the shared object files of the OpenBLAS library version compiled and installed in /opt.

You must be the operating system administrator to use this library. Therefore, do not attempt to use it without telling your system administrator. If you have the ROOT password, you will be responsible for everything you do on your operating system. Other details can also be found here.

Note

You do not have to use the ropenblas() function in every R session. Once the function is used, R will always consider using the OpenBLAS library in future sessions.

See also

Author

Pedro Rafael D. Marinho (e-mail: pedro.rafael.marinho@gmail.com)

Examples

# ropenblas()