Compile a version of R on GNU/Linux systems

This function is responsible for compiling a version of the R language.

rcompiler(x = NULL, version_openblas = NULL)

Arguments

x
Version of R you want to compile. By default (x = NULL) will be compiled the latest stable version of the R language. For example, x = "3.6.2" will compile and link R-3.6.2 version as the major version on your system.

version_openblas
OpenBLAS library version that will be linked to the R code that will be compiled. By default, if version_openblas = NULL, the latest stable version of the OpenBLAS library will be linked.

Details

The rcompiler() function will compile the R language and make the R and Rscript binaries available in the /usr/bin directory. Therefore, the version compiled by rcompiler() will be available to all users of the Linux system. The rcompiler() function will work on any GNU/Linux distribution, as long as your distribution has the following dependencies:

GNU Make;

GNU GCC Compiler (C and Fortran).

Note: If the above dependencies are not installed, the function will automatically identify. In the /opt/R directory subdirectories will be created with the compiled version numbering of R. For example, when running rcompiler(x = "version_r"), at the end of the language compilation process, the /opt/R/version_r directory will be created. The function will link the R and Rscript binaries from the current section of R to the respective newly compiled binaries found in /opt/R/version_r/lib64/R/bin, where, for example, version_r could be some version of R like "3.6.2" or any other version. If version_r = NULL, the latest stable version of R will be compiled.

See also

Examples

# rcompiler()